WebSocket을 이용한 실시간 채팅 기능 구현

+++ src/main/java/com/attica/athens/domain/user/dao/TempUserRepository.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TempUserRepository extends JpaRepository<TempUser, Integer> {
+public interface TempUserRepository extends JpaRepository<TempUser, Long> {
 
     Optional<TempUser> findByUuid(UUID uuid);
 }
