+++ src/main/java/com/attica/athens/domain/user/dao/UserRepository.java
@@ -5,7 +5,8 @@ import com.attica.athens.domain.user.domain.User;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
+
     Boolean existsByUsername(String username);
 
     Optional<User> findByUsername(String username);
